backend/src/auth/schemas.py
from pydantic import BaseModel, EmailStr

class RegisterSchema(BaseModel):
    name: str
    email: EmailStr
    password: str

class LoginSchema(BaseModel):
    email: EmailStr
    password: str

class TokenResponse(BaseModel):
    access_token: str
    token_type: str = "bearer"

class EmailSchema(BaseModel):
    email: str

class VerifyEmailSchema(BaseModel):
    email: str
    code: str