backend/requirements.txt
fastapi[all]
python-pptx
python-docx
pdfplumber
numpy
pandas
sentence-transformers
huggingface-hub
sentencepiece
protobuf
pydantic-settings
docx2md
pymupdf4llm
PyMuPDF
qdrant-client
openpyxl
psycopg2-binary
SQLAlchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.3.0
requests

